Trying to limit the impact that changes for MovingImages has on @schwa’s codebase.
