+++ MovingImages/MISVGUtilities.swift
@@ -1,0 +1,49 @@
+//
+//  MISVGUtilities.swift
+//  SwiftSVG
+//
+//  Created by Kevin Meaney on 21/09/2015.
+//  Copyright © 2015 No. All rights reserved.
+//
+
+import Foundation
+
+let defaultSaveFolder = "~/Desktop/Current/swiftsvg"
+
+public func jsonObjectToString(jsonObject: AnyObject) -> String? {
+    if NSJSONSerialization.isValidJSONObject(jsonObject) {
+        let data = try? NSJSONSerialization.dataWithJSONObject(jsonObject,
+            options: NSJSONWritingOptions.PrettyPrinted)
+        if let data = data,
+            let jsonString = NSString(data: data, encoding: NSUTF8StringEncoding) {
+                return jsonString as String
+        }
+    }
+    return nil
+}
+
+public func writeMovingImagesJSON(jsonObject: [NSString : AnyObject], sourceFileURL: NSURL) {
+    guard let fileName = sourceFileURL.lastPathComponent else {
+        return
+    }
+    
+    let shortName = NSString(string: fileName).stringByDeletingPathExtension
+    let newName = shortName.stringByAppendingString(".json")
+    let saveFolder = NSString(string: defaultSaveFolder).stringByExpandingTildeInPath
+    let folderURL = NSURL(fileURLWithPath: saveFolder, isDirectory: true)
+    
+    guard let newFileURL = NSURL(string: newName, relativeToURL: folderURL) else {
+        return
+    }
+    
+    guard let jsonString = jsonObjectToString(jsonObject) else {
+        return
+    }
+    
+    do {
+        try jsonString.writeToURL(newFileURL, atomically: false, encoding: NSUTF8StringEncoding)
+    }
+    catch {
+        print("Failed to save file: \(saveFolder)/\(newName)")
+    }
+}
